Fix #6045 - Parse gracefully the connection on Test Connection  (#6149)
* Parse gracefully test connection

* Test file renames
